+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+This release requires a server using üWave Core 0.5.0-alpha.8 or higher.
+
+Features:
+ * **Breaking:** Use the server-default artist/title when adding search results. (#2359)
+
 ## 2.0.0-alpha.8 / 23 May 2022
 While this is an alpha release, new servers should use this rather than an older "stable" version.